switch
Switches are just like the if statements. But instead of just using numeric, the statement uses words, phrases, and sentences. You start out with the switch or the setting where the conversations are taking place at. then you must have your cases, or your words or phrases of certain objects. Then you must print certain phrases for your specific cases. Its especially fun using a default when you have a word that relates to none of the cases. for example...
void Start () {
		switch(Coliseum){
			case "main":
				print("Welcome to the Games!");
			break;
			case "archer":
				print("Light the pyre!");
			break;
			case "saber":
				print("Wow! Behold the dance of blades!");
			break;
			case "lancer":
				print("Eeewww! That's gotta hurt!");
			break;
                        case "rider":
				print("Let the Chariot race begin!");
			break;
			case "assassin":
				print("Hey, where did he go?");
			break;
			case "berserker":
				print("NO MERCY!");
			break;
			case "tamer":
				print("Back! Back!");
			break;
                        case "Champion":
				print("Congratulations Champion!");
			break;
			default:
				print("Go, Go, Go! Win, Win, Win!");
			break;
		}
	}